// types/dto.ts
export interface BookDto {
  isbn: string;
  title: string;
  summary: string;
  /** @format date-time */
  publishedDate: string;
  titleImage?: string;
  authorList: Simple[];
  translator?: string;
  /** @format int32 */
  price?: number;
  publisher: Simple;
}

export interface Detail {
  /** @format int64 */
  id: number;
  name: string;
  engName?: string;
  logo?: string;
  isOfficial: boolean;
  description?: string;
  urls: UrlInfo[];
  books: BookDto[];
  tags: TagDto[];
}

export interface Simple {
  /** @format int64 */
  id: number;
  name: string;
}